refactoring - extractors refactoring
